huawei: use AT^SYSCFG for radio setting operations on 3G only modems
AT^SYSCFGEX must be used on LTE Huawei modems to enable LTE support.
But some modems (or firmwares?) do not support this command and AT^SYSCFG
must be used to get/set radio settings.
This has been introduced in commit:
22adf6402c828f8b8cca1b65d8a46ba7792eb787

There is a bug in this commit and AT^SYSCFGEX commands are used even on
modems not supporting it.