udevng: Make naming more clear
The input parameters and variables in add_device were named in such a
way that it was hard to keep track which parameters were specific to the
parent modem device, and which parameters were specific to the child
sub-device.  Make the naming more explicit and rework the parameter
order to be more consistent.