cbs: add support for UMTS CBS format
We recently learned that some devices (OP Nord N100 in this case) don't
translate CBS messages down to the GSM format, which resulted in errors
parsing CBS and no alerts.

Patching here because it's not the cleanest solution and we are talking
about a user actively missing CMAS Presidential Level alerts.

[ratchanan@ubports.com:
- Add patch description from packaging repo's commit message.
- Fixup whitespace according to checkpatch.pl.
]